#2 Design add main container
- HTML title 수정
- main page layout 생성
