+++ src/pages/Home/Home.css
@@ -1,0 +1,44 @@
+.container {
+    max-width: 1440px ;
+    min-height: 100vh;
+    width: 100%;
+    height: 100%;
+    margin: 0 auto;
+
+    background-color: yellowgreen;
+}
+
+.kanban_container {
+display: flex;
+flex-direction: row;
+flex-wrap: nowrap;
+justify-content: space-between;
+}
+
+/* 각각 Kanban Area */
+.kanban_container > ul {
+    max-width: 340px;
+    min-height: 100vh;
+    width: 100%;
+    height: 100%;
+    background-color: beige;
+    border: 1px solid #333;
+}
+
+.kanban_container > ul .title {
+    position: relative;
+    display: inline-block;
+    width: 100%;
+    margin: 10px 0;
+    font-size: 24px;
+    text-align: center;
+}
+
+.kanban_container > ul .title::before {
+    position: absolute;
+    bottom:0 ;
+    left: 0;
+    width: 100px;
+    height: 3px;
+    background-color: red;
+}
