+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import Home from "./Home";
+import Home from "./Home/Home";
 import Detail from "./Detail";
 import NotFound from "./NotFound";
